Make AClientConnection generic
